backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.120.4",
    "pillow>=12.0.0",
    "pydantic>=2.12.3",
    "pymongo>=4.15.3",
]
